+++ test/basepack_test_app/app/components/book_grid_with_mass_assignment_security.rb
@@ -1,0 +1,9 @@
+class BookGridWithMassAssignmentSecurity < Netzke::Basepack::GridPanel
+  def configure
+    super
+    config.model = "Book"
+
+    # Only allow assigning those attributes that are accessible for the :user role in the Book model
+    config.role = :user
+  end
+end