+++ test/basepack_test_app/app/models/book.rb
@@ -44,6 +44,8 @@ class Book < ActiveRecord::Base
   validates_presence_of :title
 
   scope :sorted_by_author_name, lambda { |dir| joins(:author).order("authors.last_name #{dir}, authors.first_name #{dir}") }
+
+  attr_protected :exemplars, :author_id, :as => :user
 end
 
 end